+++ bars.js
@@ -1,11 +1,13 @@
 /*
  * Simple Bar chart
  */
-export default function(config,api) {
+export default function(config,helper) {
 
-  function Bars(config, api) {
+  //Link Bars to the helper object in helper.js
+  var Bars = Object.create(helper);
+
+  Bars.init = function(config){
     var vm = this;
-    vm.api = api; 
 
     vm._config = config ? config : {};
     vm._data = [];
@@ -20,28 +22,28 @@ export default function(config,api) {
       }
     };
     vm._scales = {};
-    //vm._tip = d3.tip().attr('class', 'd3-tip tip-bars').html(vm._config.data.tip || function(d){ return d;});
     vm._tip = d3.tip()
       .attr('class', 'd3-tip')
       .direction('n')
       .html(vm._config.tip || function(d){ return vm._config._format(d[vm._config.y])});
-  }
+  };
+
 
   //-------------------------------
   //User config functions
-  Bars.prototype.id = function(columnName) {
+  Bars.id = function(columnName) {
     var vm = this;
     vm._config.id = columnName;
     return vm;
   }
 
-  Bars.prototype.x = function(columnName) {
+  Bars.x = function(columnName) {
     var vm = this;
     vm._config.x = columnName;
     return vm;
   }
 
-  Bars.prototype.y = function(columnName) {
+  Bars.y = function(columnName) {
     var vm = this;
     vm._config.y = columnName;
     return vm;
@@ -51,7 +53,7 @@ export default function(config,api) {
    * Used to draw a bar chart with multiple bars per group
    * @param {array} columns 
    */
-  Bars.prototype.groupBy = function(columns) {
+  Bars.groupBy = function(columns) {
     var vm = this;
     vm._config.groupBy = columns;
     return vm;
@@ -61,7 +63,7 @@ export default function(config,api) {
    * Used to draw a bar chart stacked with multiple bars per group
    * @param {array} columns 
    */
-  Bars.prototype.stackBy = function(columnName) {
+  Bars.stackBy = function(columnName) {
     var vm = this;
     vm._config.stackBy = columnName;
     return vm;
@@ -71,7 +73,7 @@ export default function(config,api) {
    * column name used for the domain values
    * @param {string} columnName 
    */
-  Bars.prototype.color = function(columnName) {
+  Bars.color = function(columnName) {
     var vm = this;
     vm._config.color = columnName;
     return vm;
@@ -81,7 +83,7 @@ export default function(config,api) {
    * array of values used 
    * @param {array or scale} columnName 
    */
-  Bars.prototype.colorScale = function(colorScale) {
+  Bars.colorScale = function(colorScale) {
     var vm = this;
     if(Array.isArray(colorScale)) {
       //Using an array of colors for the range 
@@ -94,7 +96,7 @@ export default function(config,api) {
     return vm;
   }
 
-  Bars.prototype.sortBy = function(option) {
+  Bars.sortBy = function(option) {
     //option = string [asc,desc] 
     //option = array for groupBy and stackBy
     var vm = this;
@@ -102,7 +104,7 @@ export default function(config,api) {
     return vm;
   }
 
-  Bars.prototype.format = function(format) {
+  Bars.format = function(format) {
     var vm = this;
     if (typeof format == 'function' || format instanceof Function)
       vm._config._format = format;
@@ -111,13 +113,13 @@ export default function(config,api) {
     return vm;
   }
 
-  Bars.prototype.tip = function(tip) {
+  Bars.tip = function(tip) {
     var vm = this;
     vm._config.tip = tip; 
     return vm;
   }
 
-  Bars.prototype.legend = function(legend) {
+  Bars.legend = function(legend) {
     var vm = this;
     vm._config.legend = legend; 
     return vm;
@@ -126,12 +128,12 @@ export default function(config,api) {
 
   //-------------------------------
   //Triggered by the chart.js;
-  Bars.prototype.data = function(data) {
+  Bars.data = function(data) {
     var vm = this;
   
-    if(vm._config.data.filter){
+    if(vm._config.filter){
       //In case we want to filter observations
-      data = data.filter(vm._config.data.filter);
+      data = data.filter(vm._config.filter);
     }
     
     if(vm._config.hasOwnProperty('stackBy') && Array.isArray(vm._config.stackBy) && vm._config.stackBy.length > 0 ){
@@ -155,7 +157,7 @@ export default function(config,api) {
       });
     }
 
-    if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles')){
+    if(vm._config.hasOwnProperty('quantiles')){
       vm._quantiles = vm._setQuantile(data);
       vm._minMax = d3.extent(data, function(d) { return +d[vm._config.color]; })
     }
@@ -163,7 +165,7 @@ export default function(config,api) {
     return vm;
   }
 
-  Bars.prototype.scales = function(s) {
+  Bars.scales = function(s) {
     var vm = this;
     var config; 
     //vm._scales = s;
@@ -179,18 +181,18 @@ export default function(config,api) {
       config = {
         column: vm._config.x,
         type: vm._config.xAxis.scale,
-        range: [0, vm.api.chart.width],
+        range: [0, vm.chart.width],
         minZero: vm._config.xAxis.minZero
       };
-      vm._scales.x = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.x = vm.utils.generateScale(vm._data, config);
 
       config = {
         column: vm._config.y,
         type: vm._config.yAxis.scale,
-        range: [vm.api.chart.height, 0],
+        range: [vm.chart.height, 0],
         minZero: vm._config.xAxis.minZero
       };
-      vm._scales.y = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.y = vm.utils.generateScale(vm._data, config);
     }
 
     //GroupBy bars on the xAxis
@@ -200,10 +202,10 @@ export default function(config,api) {
           column: vm._config.x,
           type: vm._config.xAxis.scale,
           groupBy: 'parent', 
-          range: [0, vm.api.chart.width],
+          range: [0, vm.chart.width],
           minZero: vm._config.xAxis.minZero
       };
-      vm._scales.x = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.x = vm.utils.generateScale(vm._data, config);
 
       /* Generate groupBy scale */
       config = {
@@ -212,18 +214,18 @@ export default function(config,api) {
           groupBy: 'children', 
           range: [0, vm._scales.x.bandwidth()],
       };
-      vm._scales.groupBy = vm.api.utils.generateScale(vm._data, config);
-      //vm.api.chart.scales.groupBy = vm._scales.groupBy; 
+      vm._scales.groupBy = vm.utils.generateScale(vm._data, config);
+      //vm.chart.scales.groupBy = vm._scales.groupBy; 
       
       /* Generate y scale */
       config = {
         column: vm._config.groupBy,
         type: vm._config.yAxis.scale,
         groupBy: 'data', 
-        range: [vm.api.chart.height, 0],
+        range: [vm.chart.height, 0],
         minZero: vm._config.xAxis.minZero
       };
-      vm._scales.y = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.y = vm.utils.generateScale(vm._data, config);
     }
 
     //GroupBy bars on the yAxis
@@ -233,10 +235,10 @@ export default function(config,api) {
           column: vm._config.y,
           type: vm._config.yAxis.scale,
           groupBy: 'parent', 
-          range: [0, vm.api.chart.height],
+          range: [0, vm.chart.height],
           minZero: vm._config.xAxis.minZero
       };
-      vm._scales.y = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.y = vm.utils.generateScale(vm._data, config);
 
       /* Generate groupBy scale */
       config = {
@@ -245,18 +247,18 @@ export default function(config,api) {
           groupBy: 'children', 
           range: [0, vm._scales.y.bandwidth()],
       };
-      vm._scales.groupBy = vm.api.utils.generateScale(vm._data, config);
-      //vm.api.chart.scales.groupBy = vm._scales.groupBy; 
+      vm._scales.groupBy = vm.utils.generateScale(vm._data, config);
+      //vm.chart.scales.groupBy = vm._scales.groupBy; 
       
       /* Generate x scale */
       config = {
         column: vm._config.groupBy,
         type: vm._config.xAxis.scale,
         groupBy: 'data', 
-        range: [0, vm.api.chart.width],
+        range: [0, vm.chart.width],
         minZero: vm._config.xAxis.minZero
       };
-      vm._scales.x = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.x = vm.utils.generateScale(vm._data, config);
     }
 
 
@@ -267,20 +269,20 @@ export default function(config,api) {
           column: vm._config.x,
           type: vm._config.xAxis.scale,
           stackBy: 'parent', 
-          range: [0, vm.api.chart.width],
+          range: [0, vm.chart.width],
           minZero: vm._config.xAxis.minZero
       };
-      vm._scales.x = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.x = vm.utils.generateScale(vm._data, config);
       
       /* Generate y scale */
       config = {
         column: '',
         stackBy: 'data',
         type: vm._config.yAxis.scale,
-        range: [vm.api.chart.height, 0],
+        range: [vm.chart.height, 0],
         minZero: vm._config.xAxis.minZero
       };
-      vm._scales.y = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.y = vm.utils.generateScale(vm._data, config);
     }
 
      //Stack bars on the yAxis
@@ -290,25 +292,25 @@ export default function(config,api) {
           column: '',
           type: vm._config.xAxis.scale,
           stackBy: 'data', 
-          range: [0, vm.api.chart.width],
+          range: [0, vm.chart.width],
           minZero: vm._config.xAxis.minZero
       };
-      vm._scales.x = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.x = vm.utils.generateScale(vm._data, config);
       
       /* Generate y scale */
       config = {
         column: vm._config.y,
         stackBy: 'parent',
         type: vm._config.yAxis.scale,
-        range: [vm.api.chart.height, 0],
+        range: [vm.chart.height, 0],
         minZero: vm._config.xAxis.minZero
       };
-      vm._scales.y = vm.api.utils.generateScale(vm._data, config);
+      vm._scales.y = vm.utils.generateScale(vm._data, config);
     }
 
 
-    //vm.api.chart.scales.x = vm._scales.x;
-    //vm.api.chart.scales.y = vm._scales.y;
+    //vm.chart.scales.x = vm._scales.x;
+    //vm.chart.scales.y = vm._scales.y;
 
     if(!vm._scales.hasOwnProperty('color')){
       if(vm._config.hasOwnProperty('colorScale'))
@@ -317,29 +319,29 @@ export default function(config,api) {
         vm._scales.color = d3.scaleOrdinal(d3.schemeCategory10);
     }
     
-    if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles'))
+    if(vm._config.hasOwnProperty('quantiles'))
       vm._scales.color = false;
     return vm;
   }
 
-  Bars.prototype.draw = function() {
+  Bars.draw = function() {
     var vm = this;
 
     if(vm._config.hasOwnProperty('groupBy') ){
-      if(vm._config.hasOwnProperty('x') ) vm.drawGroupByXAxis();
-      if(vm._config.hasOwnProperty('y') )  vm.drawGroupByYAxis();
+      if(vm._config.hasOwnProperty('x') ) vm._drawGroupByXAxis();
+      if(vm._config.hasOwnProperty('y') )  vm._drawGroupByYAxis();
       return vm; 
     }
 
     if(vm._config.hasOwnProperty('stackBy')){
-      if(vm._config.hasOwnProperty('x')) vm.drawStackByXAxis();
-      if(vm._config.hasOwnProperty('y')) vm.drawStackByYAxis();
+      if(vm._config.hasOwnProperty('x')) vm._drawStackByXAxis();
+      if(vm._config.hasOwnProperty('y')) vm._drawStackByYAxis();
       return vm; 
     } 
 
-    vm.api.chart.svg().call(vm._tip);
+    vm.chart.svg().call(vm._tip);
   
-    vm.api.chart.svg().selectAll(".bar")
+    vm.chart.svg().selectAll(".bar")
       .data(vm._data)
       .enter().append("rect")
         .attr("class", "bar")
@@ -362,103 +364,60 @@ export default function(config,api) {
           return vm._scales.x.bandwidth ? vm._scales.x.bandwidth() : vm._scales.x(d[vm._config.x]) 
         })
         .attr("height", function(d) { 
-          return vm._scales.y.bandwidth ? vm._scales.y.bandwidth() :  vm.api.chart.height - vm._scales.y(d[vm._config.y]); 
+          return vm._scales.y.bandwidth ? vm._scales.y.bandwidth() :  vm.chart.height - vm._scales.y(d[vm._config.y]); 
         })
         .attr("fill", function(d){             
           return vm._scales.color !== false ? vm._scales.color(d[vm._config.color]): vm._getQuantileColor(d[vm._config.color],'default');
         })
         .style("opacity", 0.9)
         .on('mouseover', function(d,i) {
-          if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles') && vm._config.bars.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
+          if(vm._config.hasOwnProperty('quantiles') && vm._config.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
             d3.select(this).attr('fill', function(d) {
                 return vm._getQuantileColor(d[vm._config.color],'onHover');
             })
           }
           vm._tip.show(d, d3.select(this).node());
 
-          if (vm._config.data.onmouseover) { //External function call, must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseover.call(this, d, i);
+          if (vm._config.hasOwnProperty('onmouseover')) { //External function call, must be after all the internal code; allowing the user to overide 
+            vm._config.onmouseover.call(this, d, i);
           }
 
         })
         .on('mouseout', function(d,i) {
-          if(vm._config.bars && vm._config.bars.quantiles.colorsOnHover){ //OnHover reset default color
+          if(vm._config.quantiles.colorsOnHover){ //OnHover reset default color
             d3.select(this).attr('fill', function(d) {
               return vm._getQuantileColor(d[vm._config.color],'default');
             })
           }
           vm._tip.hide();
 
-          if (vm._config.data.onmouseout) { //External function call, must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseout.call(this, d, i)
+          if (vm._config.hasOwnProperty('onmouseout')) { //External function call, must be after all the internal code; allowing the user to overide 
+            vm._config.onmouseout.call(this, d, i)
           }
         })
         .on('click', function(d,i) {
-          if (vm._config.data.onclick) {
-            vm._config.data.onclick.call(this, d, i)
+          if (vm._config.hasOwnProperty('click')) {
+            vm._config.onclick.call(this, d, i)
           }
         });
 
-    
-    
-    
-    
-    /*if(vm._config.xAxis.enabled) {
-       vm.api.chart.svg().append("g")
-          .attr("class", "xAxis axis")
-          .attr("transform", "translate(0," + vm.api.chart.height + ")")
-          .call(d3.axisBottom(vm._scales.x)
-            .tickValues(vm._config.xAxis.tickValues)
-            .tickFormat(vm._config.xAxis.tickFormat)
-          );
-        //vm.api.chart.svg().selectAll(".xAxis.axis text").attr("transform", "translate(0,10)rotate(-20)");
-    }*/
-
-    /*if(vm._config.yAxis.enabled) {
-      if(vm._config.yAxis.position == 'right') {
-        var yAxis = d3.axisRight(vm._scales.y)
-              .ticks(vm._config.yAxis.ticks)
-              .tickValues(vm._config.yAxis.tickValues)
-              .tickFormat(vm._config.yAxis.tickFormat);
-      } else {
-        var yAxis = d3.axisLeft(vm._scales.y)
-              .ticks(vm._config.yAxis.ticks)
-              .tickValues(vm._config.yAxis.tickValues)
-              .tickFormat(vm._config.yAxis.tickFormat);
-      }
-      var axisY = vm.api.chart.svg().append("g")
-          .attr("class", "yAxis axis");
-      if(vm._config.yAxis.position == 'right')
-        axisY.attr("transform", "translate(" + vm.api.chart.width + ",0)");
-      axisY.call(yAxis);
-        /*
-        Axis Title
-        .append("text")
-          .attr("transform", "rotate(-90)")
-          .attr("y", 6)
-          .attr("dy", "0.71em")
-          .attr("text-anchor", "end")
-          .text("Frequency");
-        
-    }*/
-    
-    
-    
     return vm;
   } 
 
+
+  //
   /** 
    * Draw bars grouped by 
   */
-  Bars.prototype.drawGroupByXAxis = function(){
+  Bars._drawGroupByXAxis = function(){
     var vm = this;
     vm._tip.html(vm._config.tip || function(d){ 
       return d.key + '<br>' +vm._config._format(d.value)
     });
 
-    vm.api.chart.svg().call(vm._tip);
+    vm.chart.svg().call(vm._tip);
     
-    vm.api.chart.svg().append("g")
+    vm.chart.svg().append("g")
       .selectAll("g")
       .data(vm._data)
       .enter().append("g")
@@ -471,52 +430,50 @@ export default function(config,api) {
           return vm._scales.y(d.value); 
         })
         .attr("width", vm._scales.groupBy.bandwidth())
-        .attr("height", function(d) { return vm.api.chart.height - vm._scales.y(d.value); })
+        .attr("height", function(d) { return vm.chart.height - vm._scales.y(d.value); })
         .attr("fill", function(d) { return vm._scales.color(d.key); }) 
         .on('mouseover', function(d,i) {
-          if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles') && vm._config.bars.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
+          if(vm._config.hasOwnProperty('quantiles') && vm._config.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
             d3.select(this).attr('fill', function(d) {
                 return vm._getQuantileColor(d[vm._config.color],'onHover');
             })
           }
           vm._tip.show(d, d3.select(this).node());
 
-          if (vm._config.data.onmouseover) { 
+          if (vm._config.hasOwnProperty('onmouseover')) { 
             //External function call. It must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseover.call(this, d, i);
+            vm._config.onmouseover.call(this, d, i);
           }
 
         })
         .on('mouseout', function(d,i) {
-          if(vm._config.bars && vm._config.bars.quantiles.colorsOnHover){ //OnHover reset default color
+          if(vm._config.hasOwnProperty('quantiles') && vm._config.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
             d3.select(this).attr('fill', function(d) {
               return vm._getQuantileColor(d[vm._config.color],'default');
             })
           }
           vm._tip.hide();
 
-          if (vm._config.data.onmouseout) { //External function call, must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseout.call(this, d, i)
+          if (vm._config.hasOwnProperty('onmouseout')) { //External function call, must be after all the internal code; allowing the user to overide 
+            vm._config.onmouseout.call(this, d, i)
           }
         })
         .on('click', function(d,i) {
-          if (vm._config.data.onclick) {
-            vm._config.data.onclick.call(this, d, i)
+          if (vm._config.hasOwnProperty('click')) {
+            vm._config.onclick.call(this, d, i)
           }
         });
-
-   
   }
 
-  Bars.prototype.drawGroupByYAxis = function(){
+  Bars._drawGroupByYAxis = function(){
     var vm = this;
     vm._tip.html(vm._config.tip || function(d){ 
       return d.key + '<br>' +vm._config._format(d.value)
     });
 
-    vm.api.chart.svg().call(vm._tip);
+    vm.chart.svg().call(vm._tip);
     
-    vm.api.chart.svg().append("g")
+    vm.chart.svg().append("g")
       .selectAll("g")
       .data(vm._data)
       .enter().append("g")
@@ -530,40 +487,39 @@ export default function(config,api) {
         .attr("height", vm._scales.groupBy.bandwidth())
         .attr("fill", function(d) { return vm._scales.color(d.key); }) 
         .on('mouseover', function(d,i) {
-          if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles') && vm._config.bars.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
+          if(vm._config.hasOwnProperty('quantiles') && vm._config.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
             d3.select(this).attr('fill', function(d) {
                 return vm._getQuantileColor(d[vm._config.color],'onHover');
             })
           }
           vm._tip.show(d, d3.select(this).node());
 
-          if (vm._config.data.onmouseover) { 
+          if (vm._config.hasOwnProperty('onmouseover')) { 
             //External function call. It must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseover.call(this, d, i);
+            vm._config.onmouseover.call(this, d, i);
           }
 
         })
         .on('mouseout', function(d,i) {
-          if(vm._config.bars && vm._config.bars.quantiles.colorsOnHover){ //OnHover reset default color
+          if(vm._config.quantiles.colorsOnHover){ //OnHover reset default color
             d3.select(this).attr('fill', function(d) {
               return vm._getQuantileColor(d[vm._config.color],'default');
             })
           }
           vm._tip.hide();
 
-          if (vm._config.data.onmouseout) { //External function call, must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseout.call(this, d, i)
+          if (vm._config.hasOwnProperty('onmouseout')) { //External function call, must be after all the internal code; allowing the user to overide 
+            vm._config.onmouseout.call(this, d, i)
           }
         })
         .on('click', function(d,i) {
-          if (vm._config.data.onclick) {
-            vm._config.data.onclick.call(this, d, i)
+          if (vm._config.hasOwnProperty('click')) {
+            vm._config.onclick.call(this, d, i)
           }
         });
-
   }
 
-  Bars.prototype.drawStackByXAxis = function(){
+  Bars._drawStackByXAxis = function(){
     var vm = this;
     vm._tip.html(vm._config.tip || function(d){ 
       var cat = ''
@@ -575,9 +531,9 @@ export default function(config,api) {
       return cat + '<br>' +vm._config._format(d[1]-d[0])
     });
 
-    vm.api.chart.svg().call(vm._tip);
+    vm.chart.svg().call(vm._tip);
 
-    vm.api.chart.svg().append("g")
+    vm.chart.svg().append("g")
       .selectAll("g")
       .data(vm._data)
       .enter().append("g")
@@ -591,41 +547,41 @@ export default function(config,api) {
         .attr("width", function(d){ return vm._scales.x.bandwidth()})
         .attr("height", function(d) { return vm._scales.y(d[0]) - vm._scales.y(d[1]); })
         .on('mouseover', function(d,i) {
-          if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles') && vm._config.bars.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
+          if(vm._config.hasOwnProperty('quantiles') && vm._config.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
             d3.select(this).attr('fill', function(d) {
                 return vm._getQuantileColor(d[vm._config.color],'onHover');
             })
           }
           vm._tip.show(d, d3.select(this).node());
 
-          if (vm._config.data.onmouseover) { 
+          if (vm._config.hasOwnProperty('onmouseover')) { 
             //External function call. It must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseover.call(this, d, i);
+            vm._config.onmouseover.call(this, d, i);
           }
 
         })
         .on('mouseout', function(d,i) {
-          if(vm._config.bars && vm._config.bars.quantiles.colorsOnHover){ //OnHover reset default color
+          if(vm._config.quantiles.colorsOnHover){ //OnHover reset default color
             d3.select(this).attr('fill', function(d) {
               return vm._getQuantileColor(d[vm._config.color],'default');
             })
           }
           vm._tip.hide();
 
-          if (vm._config.data.onmouseout) { //External function call, must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseout.call(this, d, i)
+          if (vm._config.hasOwnProperty('onmouseout')) { //External function call, must be after all the internal code; allowing the user to overide 
+            vm._config.onmouseout.call(this, d, i)
           }
         })
         .on('click', function(d,i) {
-          if (vm._config.data.onclick) {
-            vm._config.data.onclick.call(this, d, i)
+          if (vm._config.hasOwnProperty('click')) {
+            vm._config.onclick.call(this, d, i)
           }
         });
         
 
   }
 
-  Bars.prototype.drawStackByYAxis = function(){
+  Bars._drawStackByYAxis = function(){
     var vm = this;
 
     vm._tip.html(vm._config.tip || function(d){ 
@@ -638,9 +594,9 @@ export default function(config,api) {
       return cat + '<br>' +vm._config._format(d[1]-d[0])
     });
 
-    vm.api.chart.svg().call(vm._tip);
+    vm.chart.svg().call(vm._tip);
 
-    vm.api.chart.svg().append("g")
+    vm.chart.svg().append("g")
       .selectAll("g")
       .data(vm._data)
       .enter().append("g")
@@ -653,47 +609,47 @@ export default function(config,api) {
         .attr("height", function(d){ return vm._scales.y.bandwidth()})
         .attr("width", function(d) { return vm._scales.x(d[1]) - vm._scales.x(d[0]); })
         .on('mouseover', function(d,i) {
-          if(vm._config.hasOwnProperty('bars') && vm._config.bars.hasOwnProperty('quantiles') && vm._config.bars.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
+          if(vm._config.hasOwnProperty('quantiles') && vm._config.quantiles.hasOwnProperty('colorsOnHover')){ //OnHover colors
             d3.select(this).attr('fill', function(d) {
                 return vm._getQuantileColor(d[vm._config.color],'onHover');
             })
           }
           vm._tip.show(d, d3.select(this).node());
 
-          if (vm._config.data.onmouseover) { 
+          if (vm._config.hasOwnProperty('onmouseover')) { 
             //External function call. It must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseover.call(this, d, i);
+            vm._config.onmouseover.call(this, d, i);
           }
 
         })
         .on('mouseout', function(d,i) {
-          if(vm._config.bars && vm._config.bars.quantiles.colorsOnHover){ //OnHover reset default color
+          if(vm._config.quantiles.colorsOnHover){ //OnHover reset default color
             d3.select(this).attr('fill', function(d) {
               return vm._getQuantileColor(d[vm._config.color],'default');
             })
           }
           vm._tip.hide();
 
-          if (vm._config.data.onmouseout) { //External function call, must be after all the internal code; allowing the user to overide 
-            vm._config.data.onmouseout.call(this, d, i)
+          if (vm._config.hasOwnProperty('onmouseout')) { //External function call, must be after all the internal code; allowing the user to overide 
+            vm._config.onmouseout.call(this, d, i)
           }
         })
         .on('click', function(d,i) {
-          if (vm._config.data.onclick) {
-            vm._config.data.onclick.call(this, d, i)
+          if (vm._config.hasOwnProperty('click')) {
+            vm._config.onclick.call(this, d, i)
           }
         });
         
   }
 
-  Bars.prototype._setQuantile = function(data){
+  Bars._setQuantile = function(data){
     var vm = this; 
     var values = [];
     var quantile = []; 
 
-    if(vm._config.bars && vm._config.bars.quantiles &&  vm._config.bars.quantiles.predefinedQuantiles 
-        && vm._config.bars.quantiles.predefinedQuantiles.length > 0){
-      return vm._config.bars.quantiles.predefinedQuantiles;
+    if(vm._config.quantiles &&  vm._config.quantiles.predefinedQuantiles 
+        && vm._config.quantiles.predefinedQuantiles.length > 0){
+      return vm._config.quantiles.predefinedQuantiles;
     }
 
     data.forEach(function(d){      
@@ -703,23 +659,23 @@ export default function(config,api) {
     values.sort(d3.ascending);
     
     //@TODO use quantile scale instead of manual calculations 
-    if(vm._config && vm._config.bars && vm._config.bars.quantiles && vm._config.bars.quantiles.buckets){
+    if(vm._config && vm._config.quantiles && vm._config.quantiles.buckets){
 
-      if(vm._config.bars.quantiles.ignoreZeros === true){
+      if(vm._config.quantiles.ignoreZeros === true){
         var aux = _.dropWhile(values, function(o) { return o <= 0 });
         //aux.unshift(values[0]);  
 
         quantile.push(values[0]);
         quantile.push(0);
         
-        for(var i = 1; i <= vm._config.bars.quantiles.buckets - 1; i++ ){        
-          quantile.push( d3.quantile(aux,  i* 1/(vm._config.bars.quantiles.buckets - 1) ) )
+        for(var i = 1; i <= vm._config.quantiles.buckets - 1; i++ ){        
+          quantile.push( d3.quantile(aux,  i* 1/(vm._config.quantiles.buckets - 1) ) )
         }
 
       }else{
         quantile.push( d3.quantile(values, 0) )
-        for(var i = 1; i <= vm._config.bars.quantiles.buckets; i++ ){        
-          quantile.push( d3.quantile(values,  i* 1/vm._config.bars.quantiles.buckets ) )
+        for(var i = 1; i <= vm._config.quantiles.buckets; i++ ){        
+          quantile.push( d3.quantile(values,  i* 1/vm._config.quantiles.buckets ) )
         }
       }
         
@@ -728,7 +684,7 @@ export default function(config,api) {
     }
    
     //@TODO - VALIDATE WHEN ZEROS NEED TO BE PUT ON QUANTILE 1 AND RECALCULATE NON ZERO VALUES INTO THE REST OF THE BUCKETS
-    if(vm._config.bars && vm._config.bars.quantiles && vm._config.bars.quantiles.buckets && vm._config.bars.quantiles.buckets === 5){
+    if(vm._config.quantiles && vm._config.quantiles.buckets && vm._config.quantiles.buckets === 5){
 
       if( quantile[1] === quantile[2] && quantile[2] === quantile[3] && quantile[3] === quantile[4] && quantile[4] === quantile[5]){
         quantile = [ d3.quantile(values, 0), d3.quantile(values, 0.2) ];
@@ -738,15 +694,15 @@ export default function(config,api) {
     return quantile;
   }
 
-  Bars.prototype._getQuantileColor = function(d,type){
+  Bars._getQuantileColor = function(d,type){
     var vm = this; 
     var total = parseFloat(d);
 
     //@TODO use quantile scale instead of manual calculations 
-    if(vm._config && vm._config.bars && vm._config.bars.quantiles && vm._config.bars.quantiles.colors){
+    if(vm._config && vm._config.bars.quantiles && vm._config.bars.quantiles.colors){
       if(vm._quantiles.length > 2){
 
-        if(vm._config && vm._config.bars && vm._config.bars.min !== undefined && vm._config.bars.max !== undefined){
+        if(vm._config && vm._config.bars.min !== undefined && vm._config.bars.max !== undefined){
           if(total < vm._config.bars.min || total > vm._config.bars.max){
             console.log('outOfRangeColor', total, vm._config.bars.min ,vm._config.bars.max)
             return vm._config.bars.quantiles.outOfRangeColor; 
@@ -766,23 +722,23 @@ export default function(config,api) {
           }else if(total <= vm._quantiles[3]){
             return vm._config.bars.quantiles.colors[2];//"#c20216";
           }else if(total <= vm._quantiles[4]){
-            return vm._config.bars.quantiles.colors[3];//"#750000";
+            return vm._config.quantiles.colors[3];//"#750000";
           }else if(total <= vm._quantiles[5]){
-            return vm._config.bars.quantiles.colors[4];//"#480000";
+            return vm._config.quantiles.colors[4];//"#480000";
           }
         }
 
-        if(type == 'onHover' && vm._config.bars.quantiles.colorsOnHover){
+        if(type == 'onHover' && vm._config.quantiles.colorsOnHover){
           if(total <= vm._quantiles[1]){
-            return vm._config.bars.quantiles.colorsOnHover[0];//"#f7c7c5";
+            return vm._config.quantiles.colorsOnHover[0];//"#f7c7c5";
           }else if(total <= vm._quantiles[2]){
-            return vm._config.bars.quantiles.colorsOnHover[1];//"#e65158";
+            return vm._config.quantiles.colorsOnHover[1];//"#e65158";
           }else if(total <= vm._quantiles[3]){
-            return vm._config.bars.quantiles.colorsOnHover[2];//"#c20216";
+            return vm._config.quantiles.colorsOnHover[2];//"#c20216";
           }else if(total <= vm._quantiles[4]){
-            return vm._config.bars.quantiles.colorsOnHover[3];//"#750000";
+            return vm._config.quantiles.colorsOnHover[3];//"#750000";
           }else if(total <= vm._quantiles[5]){
-            return vm._config.bars.quantiles.colorsOnHover[4];//"#480000";
+            return vm._config.quantiles.colorsOnHover[4];//"#480000";
           }
         }
 
@@ -796,11 +752,12 @@ export default function(config,api) {
         return d4theme.colors.quantiles[1];//return "#f7c7c5";
       }*/
       if(total <= vm._quantiles[1]){
-        return vm._config.bars.quantiles.colors[0];//"#f7c7c5";
+        return vm._config.quantiles.colors[0];//"#f7c7c5";
       }
     }
 
   }
 
-  return new Bars(config,api);
+  Bars.init(config);
+  return Bars;
 }