+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.0.3
 
 ## @upgrades
-* added parameter api - resusable functions from chart.js
+* Implemented OLOO - Object linked to another object with the helper object 
 * added **.groupBy()** support 
 * added **.stackBy()** support 
 * added **.tip()** support
@@ -13,7 +13,17 @@
 * added **mouseover**, **mouseout**, **click**
 * use of *.hasOwnProperty()* to verify properties
 
+## @renamed
+* vm_config.data.mouseover => vm._config.mouseover
+* vm_config.data.mouseout => vm._config.mouseout
+* vm_config.data.click => vm._config.click
+* vm_config.bars.quantiles => vm._config.quantiles
+
+
 ## @deprecated 
 * removed class 'tip-treemap'
 * vm._axes = {};
 * Bars.prototype.end 
+
+
+